Added basic Bank tests
